backend/requirements.txt
# Computer Vision & ML
opencv-python==4.8.1.78
mediapipe==0.10.9
numpy==1.24.3
tensorflow==2.15.0
scikit-learn==1.3.0
scipy==1.11.3
faiss-cpu==1.7.4
seaborn==0.13.2
albumentations==1.3.1
imbalanced-learn==0.11.0

# Backend & Web
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.4
python-socketio==5.9.0
eventlet==0.33.3

# Utilities
PyYAML
pyserial>=3.5
joblib==1.3.1
matplotlib==3.7.2
kagglehub==0.2.6
requests==2.31.0
Pillow==10.0.0

# Testing
pytest==7.4.0

#CERTIFICATION
pyopenssl==23.3.0